Update torch_dtype to torch.float16 in image_generator.py

+++ app/image-gen/image_generator.py
@@ -7,16 +7,16 @@ app = FastAPI()
 
 # load both base & refiner
 base = DiffusionPipeline.from_pretrained(
-    "stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float32, variant="fp32", use_safetensors=True
+    "stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float16, variant="fp16", use_safetensors=True
 )
 # base.to("cuda")
 refiner = DiffusionPipeline.from_pretrained(
     "stabilityai/stable-diffusion-xl-refiner-1.0",
     text_encoder_2=base.text_encoder_2,
     vae=base.vae,
-    torch_dtype=torch.float32,
+    torch_dtype=torch.float16,
     use_safetensors=True,
-    variant="fp32",
+    variant="fp16",
 )
 # refiner.to("cuda")
